Add `S3Store` and `NoStore` store classes for news.
The `S3Store` gets and puts the news JSON file to/from S3.  The
`NoStore` class is a dummy for use during development.  Added a
`--no-store` option to the extractor.

In the make file, combined the `create` and `start` targets into `run`,
since `docker start` doesn't allow the setting of environment variables.

Added the `$(TMP)/.env` target to create a file for the environment
variables for AWS keys.