In news, be consistent in how items are ordered.
In the core case, news items are presented in reverse chronological
order, with new items first and older items last, but in the `News`
class, new items were appended to the end of the list, requiring reverse
iteration.

Change `News` to match the core case.  In `__iter__()`, return a forward
iterator over items instead of a reverse one.

Replace `__iadd__()` with `add_new()`, which inserts new items at the
head of the items list.  Rename `prune()` to `remove_old()` to match.