Add class `Source` to news.
In the `Item` class, change the `source` attribute from a `URL` to a
`Source` object.  In the `Source.decode()` method, enable decoding both
an URL string and a Source dict to enable reading in the previous news
JSON format.

In `News`, change the JSON key for the Items list from "stories" to
"items".

Update the home template to use the Source `site_id` when the Source URL
is different from the Item URL.