Make news cache location default to the production value.
Add two constants to the `Cache` class: `DEFAULT_DIR` and `NEWS_FILE`.

In the extractor, replace the `--cache-path` option with `--cache-dir`,
using `Cache.DEFAULT_DIR` as the default value for the option.  Rename
the `cache` variable to `news_cache` and build its location using the
cache dir and `Cache.NEWS_FILE`.

In the server, replace the `NEWS_PATH` environment variable with the
`CACHE_DIR` environment variable, defaulting to `Cache.DEFAULT_DIR`.
Rename the `cache` variable to `news_cache` and build its location using
the cache dir and `Cache.NEWS_FILE`.