Add `lifetime` and `expired` attributes to `News`.
And remove the `cutoff` parameter from the `remove_old()` method, so
that the `News` class knows which old items to discard.  Include
`expired` in the `news.json` file.  Remove the old check for the JSON
"stories" key (which is now "items").

Remove the `--cutoff-days` option from the extractor script.

Make the etag/last-modified/304 logs "debug" instead of "info".