Implement `Reddit` site for news.
Add `Reddit` subclass of `Site`.  Update the `__init__()` methods of all
the `Site` subclasses so they uniformly take an `options` dict
parameter, and update `Sites` to pass the options to each site.

Add helper functions for the `Reddit` class to extract links from the
Reddit RSS feed `<content>` tag.

To support the Reddit private RSS feed URL, add support to the extractor
`parse_options()` function to read `REDDIT_PRIVATE_RSS_FEED` from the
environment.  Update the makefile to read this value from the new
`secrets/` directory and add it to the `.env` file for the Docker
container.