Don't propogate options to each `Feed` subclass.
Remove the `options` parameter to `Feed()` and the corresponding
`options` attribute.  Also drop `options` from all the `Feed`
subclasses.  Change the signature of `Reddit()` to require a `URL` since
the Reddit feed URL needs to be injected from outside the app.

Change `Feeds.all()` to grab the Reddit URL out of the options when
creating the `Reddit` feed object.

Additionally, change `Feeds.from_json()` to drop the check for the JSON
string parameter `s` to be consistent with `News.from_json()`, and
instead change `CachedFeeds` to build an empty `Feeds` JSON string if
the cache is missing, to be consistent with `CachedNews`.

Allow `Feeds()` to be called without the `feeds` parameter.

Fix the typing on the `items` parameter to `News()`.