Rename `dev-requirements.txt` to `requirements-dev.txt`.
This will place this file next to `requirements.txt` when sorted in
alphabetical order.

Also update the "Development Setup" section of the readme, adding `aws`,
`docker`, `jq` and `make` as required tools.  Drop the need to install
requirements and update the `debug` and `extract` makefile targets to
depend on `pip-install-requirements.stamp.txt`.

Remove the unneeded `deployment` target.