Add search box to the news page.
Add a search form to the `<header>` of the news page.  Move the page
info, page controls and search form into a `<nav>` tag inside the
`<header>` to give these page controls a common parent.

Also add a `<nav>` to the `<footer>` and put the lower page info and
page controls into it.  Swap the order of page info and page control
`<p>` tags in the footer to give them a more logical document ordering,
with the page controls both upper and lower directly adjacent to the
main content.

Update the CSS in `page.html` to give the new search form a nice layout
and to remove redundant styles.