Absolutely position the search box for all screen sizes.
Always absolutely position the search box in the upper right hand corner
of the header.  Move the search box `<form>` tag out of the `<nav>` into
the `<header>` tag after the `<h1>`.

Modify the top and bottom page controls so that they don't display when
their respective links are missing.