The server now uses `CachedNews`.
Replace direct use of `CachedFile` in the `server` package with the
`CachedNews` class, which eliminates reparsing of the news JSON if the
underlying cached JSON file hasn't changed.

Update view functions and "page" classes to use `CachedNews`.

Add `is_invalid()` function to `CachedFile` so that `CachedNews` can
check the file's validity.