Change `Item.url` from `URL` to `NormalizedURL`.
Make the normalization of an `Item` URL explicit by requiring a
`NormalizedURL` object for the `url` attribute.

Change the `Item.decode()` method to use `NormalizedURL` instead of
`URL` when decoding.

Change the `Feed.parse_entry()` method and various overrides in `Feed`
subclasses to use `NormalizedURL` instead of `URL`.

Remove the `Item` tests that check normalization of the `url` attribute
since they're now redundant.