Write mypy results to `test/mypy.txt`.
Change the `check` makefile target to write mypy results to the file
`test/mypy.txt` and check this in to the project to track progress
towards building cleanly with mypy.

Make `test/mypy.txt` depend on running tests first.

Add missing files `normalized_url.py` and corresponding test to the list
of sources in the makefile.