Use shell `find` to build makefile file list variables.
Change the `container_files` variable to use `find` to get all the files
in the `./container/` subdirectory.  Remove the explicit
`container/Dockerfile` prerequisite from the `docker-build.stamp.txt`
target, since it's now part of `container_files`.

Change the `python_files` variable to use `find` to get all the files in
the `./src/` subdirectory.