Fix mypy type import errors.
Add a `.mypy.ini` file and configure mypy to ignore missing type
annotations for feedparser.

Add type packages to the dev requirements file for `boto3`, `botocore`,
`bs4` and `requests`.

Update the makefile so that the `test/mypy.txt` file depends on the
`.mypy.ini` file and the dev requirements stamp file.