Make `Feed.is_entry_recent()` check for future published date.
Log a message when checking feed entries if the entry has a published
date that is in the future, which will cause an old entry to be re-added
to news after it expires and is removed, since news `Item` objects use
the current server time as their created time, not the feed's published
time.

Include a feed from daringfireball.net with some entries that had
published dates in the future.