Fix build error by installing Python libs with `apk`.
The container build started failing with this error from Alpine Linux:

    **error**: **externally-managed-environment**
    × This environment is externally managed
    ╰─>
        The system-wide python installation should be maintained using the system
        package manager (apk) only.

    ...

    **hint**: See PEP 668 for the detailed specification.

Change the `Dockerfile`, adding the Alpine packages to the `apk add`
command that correspond to the packages in `requirements.txt` and remove
the `pip install` command.  Also remove the `COPY` step for
`requirements.txt` since it's no longer needed in the container.

Drop `requirements.txt` from the `container_files` list in the makefile.

In `requirements.txt`, downgrade `boto3` from 1.34.10 to 1.28.12 to
match the version available from `apk`.