Accept `application/json` for news pages.
Generate a JSON response for news pages when requested.

Change the `news_page()` view function to check the `Accept` header and
return a JSON representation when appropriate.

Change the `NewsPage` class to add a `to_json()` method and to build
full URLs for the JSON representation.

Change the `Page` class to add a `to_json()` method and add the missing
`first()` method.  Remove the unused `one_page()` method.

In the `news.html` template, use the presence of the previous and next
URLs to determine the visibility of those respective navigation
sections.