Rework the `cov` makefile code coverage target.
Change the coverage file output name from the default of `.coverage` to
`coverage.sqlite`.

Move the coverage config from the `.coveragerc` file into the
`pyproject.toml` file.

Run `pytest` with the `--cov` option but don't produce any reports in
order to suppress noisy output that `pytest-cov` always writes to the
console.  Add two invocations of the `coverage` command to the makefile
recipe, one to write the top-level coverage percent to the console and a
second to produce the HTML coverage report.